Fix Travis to properly fail the build on go fmt and go vet.
When running go fmt/vet, they do not exit non-zero if they detect any issues,
they only print out information. This wraps go fmt/vet in a script that
captures the output to a temp file and checks if anything was written to it.
If any was, it prints it out then exits 1.